requirements.txt
streamlit
rembg
Pillow
opencv-python
numpy